# server/openapi.yaml
openapi: 3.1.0
info:
  title: Voxels public read API
  version: '1.0.0'
  description: |
    The read-only part of the Voxels API. Everything here is unauthenticated GET.

    Write routes, sign-in, admin, moderation, livekit, radio, metrics, guest
    passes and the internal /grid/* routes are deliberately not described.

    Most handlers answer with an envelope: `{"success": true, "<field>": ...}`,
    where `<field>` is plural for a list and singular for one record. A failed
    lookup answers `{"success": false}`, usually with status 400 and sometimes
    with 200, so check the flag rather than the status.
servers:
  - url: https://www.voxels.com
...
paths:
...
  /api/parcels/{id}.json:
    get:
      operationId: getParcel
      tags: [parcels]
      summary: One parcel with its build
      description: |
        The only parcel route that returns `content`, which holds the voxels, the
        palette, the tileset and every feature on the plot. A non-numeric id is a
        400. A parcel that is neither minted nor visible is a 400 as well.
      parameters:
        - $ref: '#/components/parameters/ParcelId'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  parcel:
                    $ref: '#/components/schemas/Parcel'
        '400':
          $ref: '#/components/responses/Failure'
...
  /api/parcels/{id}/query:
    get:
      operationId: refreshParcelFromChain
      tags: [parcels]
      summary: Re-read the parcel's owner from the contract, then return it
      description: |
        A read with a side effect: it asks the contract who owns the parcel and
        writes the answer back before answering. The `parcel` it returns is the
        model object, not the `/api/parcels/{id}.json` shape.
      parameters:
        - $ref: '#/components/parameters/ParcelId'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  parcel:
                    type: object
        '404':
          $ref: '#/components/responses/Failure'
  /api/parcels/by/{wallet}/query:
    get:
      operationId: refreshParcelsByWallet
      tags: [parcels]
      summary: Re-read every parcel a wallet owns
      description: |
        Reads the subgraph for every parcel the wallet owns, then re-queries the
        contract for each one as a just-to-make-sure step. Answers
        `{"success": false}` when the subgraph is unreachable or returns nothing,
        on the theory that an owner seeing no parcels is worse than an owner
        seeing a stale list.
      parameters:
        - name: wallet
          in: path
          required: true
          schema:
            type: string
      responses:
        '404':
          $ref: '#/components/responses/Failure'
...
  /api/wallet/{address}/parcels.json:
    get:
      operationId: getParcelsByOwner
      tags: [parcels]
      summary: Parcels a wallet owns
      parameters:
        - $ref: '#/components/parameters/WalletAddress'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  parcels:
                    type: array
                    items:
                      type: object
  /api/wallet/{address}/contributing-parcels.json:
    get:
      operationId: getContributingParcels
      tags: [parcels]
      summary: Parcels a wallet can build on but does not own
      parameters:
        - $ref: '#/components/parameters/WalletAddress'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  parcels:
                    type: array
                    items:
                      type: object
...
  /api/avatar/{wallet}/names:
    get:
      operationId: getAvatarNames
      tags: [avatars]
      summary: Every name this wallet holds
      description: No `success` on the happy path, just `name` and `names`.
      parameters:
        - $ref: '#/components/parameters/Wallet'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: [string, 'null']
                  names:
                    type: array
                    items:
                      type: string
        '400':
          $ref: '#/components/responses/Failure'
  /api/names/exists/{name}:
    get:
      operationId: nameExists
      tags: [avatars]
      summary: Whether a name is taken
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  exists:
                    type: boolean
...
  /api/helper/typeOfContract/{chain_identifier}/{contract}:
    get:
      operationId: getTypeOfContract
      tags: [collections]
      summary: Whether a contract is ERC721 or ERC1155
      parameters:
        - $ref: '#/components/parameters/ChainIdentifier'
        - name: contract
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  type:
                    type: [string, 'null']
        '400':
          $ref: '#/components/responses/Failure'
...
  /api/wearables/{wearable_id}/vox:
    get:
      operationId: getWearableVoxById
      tags: [wearables]
      summary: A wearable's MagicaVoxel model, by wearable uuid
      parameters:
        - name: wearable_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: ok
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/Failure'
  /api/wearables/{address}/{token}/vox:
    get:
      operationId: getWearableVoxByToken
      tags: [wearables]
      summary: A wearable's MagicaVoxel model, by contract and token id
      description: |
        Two-segment sibling of the route above. Express matches whichever arity
        the request has, so a single segment goes to `{wearable_id}` and two go
        here. The address is matched case-insensitively against the collection.
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ok
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/Failure'
  /w/{hash}/{format}:
    get:
      operationId: streamWearableByHash
      tags: [wearables]
      summary: A wearable's model by content hash
      description: |
        `format` has to be `vox` or `.vox`; every other value is a 404. The hash
        has to be longer than 39 characters or the lookup is skipped.
      parameters:
        - name: hash
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/WearableFormat'
      responses:
        '200':
          description: ok
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/Failure'
  /c/v2/{chain_identifier}/{collection_address}/{token_id}/{format}:
    get:
      operationId: streamWearableByChainInfo
      tags: [wearables]
      summary: A wearable's model by chain, contract and token id
      description: |
        Same handler as `/w/{hash}/{format}`. `token_id` is read as hex when it
        looks like hex and as decimal otherwise.
      parameters:
        - $ref: '#/components/parameters/ChainIdentifier'
        - name: collection_address
          in: path
          required: true
          schema:
            type: string
        - name: token_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/WearableFormat'
      responses:
        '200':
          description: ok
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/Failure'
        '404':
          $ref: '#/components/responses/Failure'
...
  /api/islands/{slug}.json:
    get:
      operationId: getIsland
      tags: [islands]
      summary: One island and every parcel on it
      description: |
        The slug is the island name lowercased with runs of whitespace turned
        into single hyphens, so `Origin City` is `origin-city`. `parcels` holds
        whole property rows.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  island:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      position:
                        $ref: '#/components/schemas/GeoJsonPoint'
                      parcels:
                        type: array
                        items:
                          type: object
        '400':
          $ref: '#/components/responses/Failure'
...
  /api/search:
    get:
      operationId: search
      tags: [search]
      summary: Full text search across the world
      description: |
        Reads a materialised view that mixes parcels, wearables and the rest, so
        `type` tells you what each hit is. Fifty results, no paging: `limit` and
        `page` are fixed in the handler. `q` is lowercased, stripped of percent
        signs and cut to eighty characters. An empty `q` returns an empty list
        with `success: true`.
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchResult'
components:
  parameters:
    ParcelId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    Wallet:
      name: wallet
      in: path
      required: true
      schema:
        type: string
    WalletAddress:
      name: address
      in: path
      required: true
      schema:
        type: string
...
    ChainIdentifier:
      name: chain_identifier
      in: path
      required: true
      schema:
        type: string
        enum: [eth, matic]
    WearableFormat:
      name: format
      in: path
      required: true
      schema:
        type: string
        enum: [vox, .vox]
...
  responses:
    Failure:
      description: The lookup did not land. Some handlers send this with status 200.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Failure'
  schemas:
    Failure:
      type: object
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
...
    GeoJsonPoint:
      type: object
      properties:
        type:
          type: string
          const: Point
        crs:
          type: object
        coordinates:
          type: array
          items:
            type: number
...
    Parcel:
      description: One parcel with its build. What `/api/parcels/{id}.json` returns.
      allOf:
        - $ref: '#/components/schemas/ParcelBounds'
        - type: object
          properties:
            id:
              type: integer
            token:
              type: [integer, 'null']
            name:
              type: [string, 'null']
            label:
              type: [string, 'null']
            description:
              type: [string, 'null']
            address:
              type: [string, 'null']
              description: The street address. About ten parcels have none.
            island:
              type: string
            suburb:
              type: [string, 'null']
            kind:
              type: string
              enum: [plot, inner, outer, unit, basement, asset]
            geometry:
              $ref: '#/components/schemas/ParcelGeometry'
            owner:
              $ref: '#/components/schemas/AvatarRef'
            parcel_users:
              type: [array, 'null']
              items:
                $ref: '#/components/schemas/AvatarRef'
            content:
              $ref: '#/components/schemas/ParcelContent'
            settings:
              type: [object, 'null']
            lightmap_url:
              type: [string, 'null']
            traffic_visits:
              type: integer
            distance_to_center:
              type: number
            distance_to_ocean:
              type: number
            distance_to_closest_common:
              type: number
            is_common:
              type: boolean
            minted:
              type: boolean
            visible:
              type: boolean
            updated_at:
              type: string
...